+++ shared/constants/followers/util.tsx
@@ -1,0 +1,29 @@
+import * as EngineGen from '@/actions/engine-gen-gen'
+import isEqual from 'lodash/isEqual'
+import {useCurrentUserState} from '../current-user'
+import {useFollowerState} from '../followers'
+
+export const onEngineIncoming = (action: EngineGen.Actions) => {
+  switch (action.type) {
+    case EngineGen.keybase1NotifyTrackingTrackingChanged: {
+      const {isTracking, username} = action.payload.params
+      useFollowerState.getState().dispatch.updateFollowing(username, isTracking)
+      break
+    }
+    case EngineGen.keybase1NotifyTrackingTrackingInfo: {
+      const {uid, followers: _newFollowers, followees: _newFollowing} = action.payload.params
+      if (useCurrentUserState.getState().uid !== uid) {
+        return
+      }
+      const newFollowers = new Set(_newFollowers)
+      const newFollowing = new Set(_newFollowing)
+      const {following: oldFollowing, followers: oldFollowers, dispatch} = useFollowerState.getState()
+      const following = isEqual(newFollowing, oldFollowing) ? oldFollowing : newFollowing
+      const followers = isEqual(newFollowers, oldFollowers) ? oldFollowers : newFollowers
+      dispatch.replace(followers, following)
+      break
+    }
+    default:
+  }
+}
+
