+++ shared/constants/platform-specific/index.desktop.tsx
@@ -5,14 +5,16 @@ import {useConfigState} from '../config'
 import * as ConfigConstants from '../config'
 import {useDaemonState} from '../daemon'
 import {useFSState} from '../fs'
+import {usePinentryState} from '../pinentry'
 import {useProfileState} from '../profile'
 import {useRouterState} from '../router2'
 import * as EngineGen from '@/actions/engine-gen-gen'
+import * as RemoteGen from '@/actions/remote-gen'
 import * as T from '../types'
 import InputMonitor from './input-monitor.desktop'
 import KB2 from '@/util/electron.desktop'
 import logger from '@/logger'
-import type {RPCError} from '@/util/errors'
+import {RPCError} from '@/util/errors'
 import {getEngine} from '@/engine'
 import {isLinux, isWindows} from '../platform.desktop'
 import {kbfsNotification} from './kbfs-notifications'
@@ -20,7 +22,10 @@ import {skipAppFocusActions} from '@/local-debug.desktop'
 import NotifyPopup from '@/util/notify-popup'
 import {noKBFSFailReason} from '@/constants/config/util'
 import {initSharedSubscriptions} from './shared'
+import {switchTab} from '../router2/util'
+import {storeRegistry} from '../store-registry'
 import {wrapErrors} from '@/util/debug'
+import {getSelectedConversation} from '@/constants/chat2/common'
 
 const {showMainWindow, activeChanged, requestWindowsStartService, dumpNodeLogger} = KB2.functions
 const {quitApp, exitApp, setOpenAtLogin, ctlQuit, copyToClipboard} = KB2.functions
@@ -153,39 +158,49 @@ export const initPlatformListener = () => {
   })
 
   useConfigState.subscribe((s, old) => {
-    if (s.loggedIn === old.loggedIn) return
-    s.dispatch.osNetworkStatusChanged(navigator.onLine, 'notavailable', true)
-  })
-
-  useConfigState.subscribe((s, prev) => {
-    if (s.appFocused !== prev.appFocused) {
-      maybePauseVideos()
+    if (s.loggedIn !== old.loggedIn) {
+      s.dispatch.osNetworkStatusChanged(navigator.onLine, 'notavailable', true)
     }
-  })
 
-  useDaemonState.subscribe((s, old) => {
-    if (s.handshakeVersion === old.handshakeVersion) return
-    if (!isWindows) return
-
-    const f = async () => {
-      const waitKey = 'pipeCheckFail'
-      const version = s.handshakeVersion
-      const {wait} = s.dispatch
-      wait(waitKey, version, true)
-      try {
-        logger.info('Checking RPC ownership')
-        if (KB2.functions.winCheckRPCOwnership) {
-          await KB2.functions.winCheckRPCOwnership()
-        }
-        wait(waitKey, version, false)
-      } catch (error_) {
-        // error will be logged in bootstrap check
-        getEngine().reset()
-        const error = error_ as RPCError
-        wait(waitKey, version, false, error.message || 'windows pipe owner fail', true)
+    if (s.appFocused !== old.appFocused) {
+      maybePauseVideos()
+      if (!old.appFocused && s.appFocused) {
+        const {dispatch} = storeRegistry.getConvoState(getSelectedConversation())
+        dispatch.loadMoreMessages({reason: 'foregrounding'})
+        dispatch.markThreadAsRead()
       }
     }
-    ignorePromise(f())
+
+    if (s.openAtLogin !== old.openAtLogin) {
+      const {openAtLogin} = s
+      const f = async () => {
+        if (__DEV__) {
+          console.log('onSetOpenAtLogin disabled for dev mode')
+          return
+        } else {
+          await T.RPCGen.configGuiSetValueRpcPromise({
+            path: ConfigConstants.openAtLoginKey,
+            value: {b: openAtLogin, isNull: false},
+          })
+        }
+        if (isLinux || isWindows) {
+          const enabled =
+            (await T.RPCGen.ctlGetOnLoginStartupRpcPromise()) === T.RPCGen.OnLoginStartupStatus.enabled
+          if (enabled !== openAtLogin) {
+            try {
+              await T.RPCGen.ctlSetOnLoginStartupRpcPromise({enabled: openAtLogin})
+            } catch (error_) {
+              const error = error_ as RPCError
+              logger.warn(`Error in sending ctlSetOnLoginStartup: ${error.message}`)
+            }
+          }
+        } else {
+          logger.info(`Login item settings changed! now ${openAtLogin ? 'on' : 'off'}`)
+          await setOpenAtLogin?.(openAtLogin)
+        }
+      }
+      ignorePromise(f())
+    }
   })
 
   const handleWindowFocusEvents = () => {
@@ -211,46 +226,39 @@ export const initPlatformListener = () => {
   }
   setupReachabilityWatcher()
 
-  useConfigState.subscribe((s, old) => {
-    if (s.openAtLogin === old.openAtLogin) return
-    const {openAtLogin} = s
-    const f = async () => {
-      if (__DEV__) {
-        console.log('onSetOpenAtLogin disabled for dev mode')
-        return
-      } else {
-        await T.RPCGen.configGuiSetValueRpcPromise({
-          path: ConfigConstants.openAtLoginKey,
-          value: {b: openAtLogin, isNull: false},
-        })
-      }
-      if (isLinux || isWindows) {
-        const enabled =
-          (await T.RPCGen.ctlGetOnLoginStartupRpcPromise()) === T.RPCGen.OnLoginStartupStatus.enabled
-        if (enabled !== openAtLogin) {
-          try {
-            await T.RPCGen.ctlSetOnLoginStartupRpcPromise({enabled: openAtLogin})
-          } catch (error_) {
-            const error = error_ as RPCError
-            logger.warn(`Error in sending ctlSetOnLoginStartup: ${error.message}`)
-          }
-        }
-      } else {
-        logger.info(`Login item settings changed! now ${openAtLogin ? 'on' : 'off'}`)
-        await setOpenAtLogin?.(openAtLogin)
-      }
-    }
-    ignorePromise(f())
-  })
-
   useDaemonState.subscribe((s, old) => {
-    if (s.handshakeState === old.handshakeState || s.handshakeState !== 'done') return
-    useConfigState.getState().dispatch.setStartupDetails({
-      conversation: Chat.noConversationIDKey,
-      followUser: '',
-      link: '',
-      tab: undefined,
-    })
+    if (s.handshakeVersion !== old.handshakeVersion) {
+      if (!isWindows) return
+
+      const f = async () => {
+        const waitKey = 'pipeCheckFail'
+        const version = s.handshakeVersion
+        const {wait} = s.dispatch
+        wait(waitKey, version, true)
+        try {
+          logger.info('Checking RPC ownership')
+          if (KB2.functions.winCheckRPCOwnership) {
+            await KB2.functions.winCheckRPCOwnership()
+          }
+          wait(waitKey, version, false)
+        } catch (error_) {
+          // error will be logged in bootstrap check
+          getEngine().reset()
+          const error = error_ as RPCError
+          wait(waitKey, version, false, error.message || 'windows pipe owner fail', true)
+        }
+      }
+      ignorePromise(f())
+    }
+
+    if (s.handshakeState !== old.handshakeState && s.handshakeState === 'done') {
+      useConfigState.getState().dispatch.setStartupDetails({
+        conversation: Chat.noConversationIDKey,
+        followUser: '',
+        link: '',
+        tab: undefined,
+      })
+    }
   })
 
   if (isLinux) {
@@ -295,4 +303,165 @@ export const initPlatformListener = () => {
   initSharedSubscriptions()
 
   ignorePromise(useFSState.getState().dispatch.setupSubscriptions())
+}
+
+const updateApp = () => {
+  const f = async () => {
+    await T.RPCGen.configStartUpdateIfNeededRpcPromise()
+  }
+  ignorePromise(f())
+  // * If user choose to update:
+  //   We'd get killed and it doesn't matter what happens here.
+  // * If user hits "Ignore":
+  //   Note that we ignore the snooze here, so the state shouldn't change,
+  //   and we'd back to where we think we still need an update. So we could
+  //   have just unset the "updating" flag.However, in case server has
+  //   decided to pull out the update between last time we asked the updater
+  //   and now, we'd be in a wrong state if we didn't check with the service.
+  //   Since user has interacted with it, we still ask the service to make
+  //   sure.
+
+  useConfigState.getState().dispatch.setUpdating()
+}
+
+export const eventFromRemoteWindows = (action: RemoteGen.Actions) => {
+  switch (action.type) {
+    case RemoteGen.resetStore:
+      break
+    case RemoteGen.openChatFromWidget: {
+      useConfigState.getState().dispatch.showMain()
+      storeRegistry.getConvoState(action.payload.conversationIDKey).dispatch.navigateToThread('inboxSmall')
+      break
+    }
+    case RemoteGen.inboxRefresh: {
+      storeRegistry.getState('chat').dispatch.inboxRefresh('widgetRefresh')
+      break
+    }
+    case RemoteGen.engineConnection: {
+      if (action.payload.connected) {
+        storeRegistry.getState('engine').dispatch.onEngineConnected()
+      } else {
+        storeRegistry.getState('engine').dispatch.onEngineDisconnected()
+      }
+      break
+    }
+    case RemoteGen.switchTab: {
+      switchTab(action.payload.tab)
+      break
+    }
+    case RemoteGen.setCriticalUpdate: {
+      storeRegistry.getState('fs').dispatch.setCriticalUpdate(action.payload.critical)
+      break
+    }
+    case RemoteGen.userFileEditsLoad: {
+      storeRegistry.getState('fs').dispatch.userFileEditsLoad()
+      break
+    }
+    case RemoteGen.openFilesFromWidget: {
+      storeRegistry.getState('fs').dispatch.dynamic.openFilesFromWidgetDesktop?.(action.payload.path)
+      break
+    }
+    case RemoteGen.saltpackFileOpen: {
+      storeRegistry.getState('deeplinks').dispatch.handleSaltPackOpen(action.payload.path)
+      break
+    }
+    case RemoteGen.pinentryOnCancel: {
+      usePinentryState.getState().dispatch.dynamic.onCancel?.()
+      break
+    }
+    case RemoteGen.pinentryOnSubmit: {
+      usePinentryState.getState().dispatch.dynamic.onSubmit?.(action.payload.password)
+      break
+    }
+    case RemoteGen.openPathInSystemFileManager: {
+      storeRegistry.getState('fs').dispatch.dynamic.openPathInSystemFileManagerDesktop?.(action.payload.path)
+      break
+    }
+    case RemoteGen.unlockFoldersSubmitPaperKey: {
+      T.RPCGen.loginPaperKeySubmitRpcPromise({paperPhrase: action.payload.paperKey}, 'unlock-folders:waiting')
+        .then(() => {
+          useConfigState.getState().dispatch.openUnlockFolders([])
+        })
+        .catch((e: unknown) => {
+          if (!(e instanceof RPCError)) return
+          useConfigState.setState(s => {
+            s.unlockFoldersError = e.desc
+          })
+        })
+      break
+    }
+    case RemoteGen.closeUnlockFolders: {
+      T.RPCGen.rekeyRekeyStatusFinishRpcPromise()
+        .then(() => {})
+        .catch(() => {})
+      useConfigState.getState().dispatch.openUnlockFolders([])
+      break
+    }
+    case RemoteGen.stop: {
+      storeRegistry.getState('settings').dispatch.stop(action.payload.exitCode)
+      break
+    }
+    case RemoteGen.trackerChangeFollow: {
+      storeRegistry.getState('tracker2').dispatch.changeFollow(action.payload.guiID, action.payload.follow)
+      break
+    }
+    case RemoteGen.trackerIgnore: {
+      storeRegistry.getState('tracker2').dispatch.ignore(action.payload.guiID)
+      break
+    }
+    case RemoteGen.trackerCloseTracker: {
+      storeRegistry.getState('tracker2').dispatch.closeTracker(action.payload.guiID)
+      break
+    }
+    case RemoteGen.trackerLoad: {
+      storeRegistry.getState('tracker2').dispatch.load(action.payload)
+      break
+    }
+    case RemoteGen.link:
+      {
+        const {link} = action.payload
+        storeRegistry.getState('deeplinks').dispatch.handleAppLink(link)
+      }
+      break
+    case RemoteGen.installerRan:
+      useConfigState.getState().dispatch.installerRan()
+      break
+    case RemoteGen.updateNow:
+      updateApp()
+      break
+    case RemoteGen.powerMonitorEvent:
+      useConfigState.getState().dispatch.powerMonitorEvent(action.payload.event)
+      break
+    case RemoteGen.showMain:
+      useConfigState.getState().dispatch.showMain()
+      break
+    case RemoteGen.dumpLogs:
+      ignorePromise(useConfigState.getState().dispatch.dumpLogs(action.payload.reason))
+      break
+    case RemoteGen.remoteWindowWantsProps:
+      useConfigState
+        .getState()
+        .dispatch.remoteWindowNeedsProps(action.payload.component, action.payload.param)
+      break
+    case RemoteGen.updateWindowMaxState:
+      useConfigState.setState(s => {
+        s.windowState.isMaximized = action.payload.max
+      })
+      break
+    case RemoteGen.updateWindowState:
+      useConfigState.getState().dispatch.updateWindowState(action.payload.windowState)
+      break
+    case RemoteGen.updateWindowShown: {
+      const win = action.payload.component
+      useConfigState.setState(s => {
+        s.windowShownCount.set(win, (s.windowShownCount.get(win) ?? 0) + 1)
+      })
+      break
+    }
+    case RemoteGen.previewConversation:
+      storeRegistry
+        .getState('chat')
+        .dispatch.previewConversation({participants: [action.payload.participant], reason: 'tracker'})
+      break
+  }
 }
